Add files via upload
The data was obtained from Mendely with the original source from the Medical City Hospital laboratory in Iraq. The diabetes dataset we used in this group project is an open-source dataset orginate from Mendeley Data website. This dataset is collected from laboratory of Medical City Hospital and the Specializes Center for Endocrinology and Diabetes-Al-Kindy Teaching Hospital to study the prediction of diabetes. All of the respndents are from Iraqi society. One target variable and several medical predictor variables are included in this dataset:

Gender is refer to characteristics of female or male of the patient.
Age is refer to the patient’s age in years.
Urea is refer the concentration measured in milligrams per deciliter.
Creatinine ratio refer to measures the amount of albumin in urine.
HbA1c is refer to average blood glucose (sugar) levels for the last two to three months.
Cholesterol (Chol) is sum of blood's cholesterol content that measures in the mmol/L.
Triglycerides(TG) is refer to most common type of fat in the body.
HDL Cholesterol is refer to “good” cholesterol that can absorbs cholesterol in the blood and carries it back to the liver.
LDL Cholesterol is refer to “bad” cholesterol, makes up most of your body's cholesterol.
VLDL Cholesterol is refer fat that produced in the liver and released into the bloodstream to supply body tissues.
BMI refer to the body mass index calculated using this formula: weight in kg/(height in m)^2).
Data Source: https://data.mendeley.com/datasets/wj9rwkp9c2/1/